Combine XCTest extensions across targets
