--- WordPress/WordPressScreenshotGeneration/XCTest+Extensions.swift
@@ -1,24 +1,0 @@
-import XCTest
-
-extension XCTestCase {
-
-    public func waitForElementToExist(element: XCUIElement, timeout: TimeInterval? = nil) {
-        let timeoutValue = timeout ?? 30
-        guard element.waitForExistence(timeout: timeoutValue) else {
-            XCTFail("Failed to find \(element) after \(timeoutValue) seconds.")
-            return
-        }
-    }
-
-    public func waitForElementToNotExist(element: XCUIElement, timeout: TimeInterval? = nil) {
-        let notExistsPredicate = NSPredicate(format: "exists == false")
-        let expectation = XCTNSPredicateExpectation(predicate: notExistsPredicate,
-                                                    object: element)
-
-        let timeoutValue = timeout ?? 30
-        guard XCTWaiter().wait(for: [expectation], timeout: timeoutValue) == .completed else {
-            XCTFail("\(element) still exists after \(timeoutValue) seconds.")
-            return
-        }
-    }
-}
